[shreegowri]adding exception to the program
