+++ .idea/modules.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ProjectModuleManager">
+    <modules>
+      <module fileurl="file://$PROJECT_DIR$/.idea/Day1_BassicAndLogicalPgm-python-.iml" filepath="$PROJECT_DIR$/.idea/Day1_BassicAndLogicalPgm-python-.iml" />
+    </modules>
+  </component>
+</project>
