--- BasicProgramingInPython/FibonacciPP.py
@@ -1,6 +1,0 @@
-a, b = 0, 1
-while a < 10:
-    print(a, end=',')#The keyword argument end can be used to avoid the newline after the output, or end the output with a different string:
-    a, b = b, a + b
-
-
