--- BasicProgramingInPython/ForLoops.py
@@ -1,3 +1,0 @@
-fruits = ['banana', 'apple', 'grapes']
-for f in fruits:
-    print(f, len(f))
