--- BasicProgramingInPython/IfElseStmt.py
@@ -1,9 +1,0 @@
-x = int(input("entre the integer value "))
-if x == 0:
-    print("Zero")
-elif x < 0:
-    print("entered negative number")
-elif x > 0:
-    print("entered the positive number")
-else:
-    print("invalid")
