--- BasicProgramingInPython/PrimeWithRange.py
@@ -1,16 +1,0 @@
-
-def prime_number_range(a_, b_):
-    for n in range(a_, b_):
-        for x in range(a_, n):
-
-            if n % x == 0:
-                print(n, 'equals', x, '*', n // x, "not prime number")
-                break
-        else:
-            print(n, 'is a prime number')
-
-
-if __name__ == '__main__':
-    a = int(input("enter the range "))
-    b = int(input("enter the range2 "))
-    prime_number_range(a, b)
