--- BasicProgramingInPython/rangeFuction.py
@@ -1,5 +1,0 @@
-# for x in range(5):
-#     print(x)
-for x in range(5, 15):
-    print(x)
-#print(sum(range(2)))
