--- BasicProgramingInPython/vowelsAndConsonents.py
@@ -1,26 +1,0 @@
-def vowels_consonants(alphabets_):
-    if alphabets_ == 'a':
-        print("vowel")
-    elif alphabets_ == 'e':
-        print("vowel")
-    elif alphabets_ == 'i':
-        print("vowel")
-    elif alphabets_ == 'o':
-        print("vowel")
-    elif alphabets_ == 'u':
-        print("vowel")
-    else:
-        print("consonants")
-
-
-if __name__ == '__main__':
-    alphabets = input("enter the alphabets ")
-    vowels_consonants(alphabets)
-
-
-
-
-
-
-
-
